separate the local cache by architecture
This is done so that the layers don't interfere with each other.
In addition, I added the 'mode=max' setting when exporting the cache so
that all layers are exported and not just the ones that end up in the
image. This is suggested if  you have the space because it allows for
future builds to use the intermediate, non-final layers during their
build.

I attempted using the registry cache but it produced a confusing
registry error - probably due to the conflicting architectures pushing
to the same location. Perhaps this can be revisted in the future but for
now I think a local cache on our single runner will work well.