remove addition of external subdirectory to ldmx-sw install
external was only ever filled by ONNX runtime being downloaded by the CMake
infrastructure. Since it was moved into the container image (v4.0), the
external subdirectory of an install will only exist if the user
specifically edits the cmake to use the older ONNX version. In this
case, they will now need to also update their local copy of .profile to
include that directory in LD_LIBRARY_PATH (if using denv) or edit the
cmake to install ONNX in one of the paths already included.